refactor(query-performance-ai): split sandbox setup out of run_campaign.py
run_campaign.py was carrying lockdown, pi-coding-agent install, pi-ai
bundle patches, workspace setup, pi process supervision, and pi-event
streaming all in one file. Pull the sandbox-side setup into a focused
sandboxing.py and the small shared utilities into _runtime.py.

- _runtime.py: CampaignError, log, atomic_write, run, plus the path
  constants both files need (PRODUCT_DIR, AUTORESEARCH_DIR, SCRIPTS_DIR).
- sandboxing.py: LockdownFailed + lockdown_network + iptables resolvers,
  install_pi_toolchain, prepare_pi_runtime, and the two pi-ai /
  pi-autoresearch bundle patch helpers.
- run_campaign.py: imports from the two above. Now 406 lines (was 758),
  scoped to coordinator preflight, workspace init, and pi orchestration.

No behavioural change. The refactor leaves room to add per-concern unit
tests for the sandboxing module separately.

Generated-By: PostHog Code
Task-Id: 9ee4b87f-fce9-4d82-aded-510477a12e4a
